+++ paimon-common/src/main/java/org/apache/paimon/fileindex/bsi/BitSliceIndexBitmapFileIndex.java
@@ -68,7 +68,7 @@ public class BitSliceIndexBitmapFileIndex implements FileIndexer {
     }
 
     @Override
-    public FileIndexReader createReader(SeekableInputStream inputStream, long start, int length) {
+    public FileIndexReader createReader(SeekableInputStream inputStream, long start, long length) {
         try {
             inputStream.seek(start);
             DataInput input = new DataInputStream(inputStream);
