+++ paimon-common/src/main/java/org/apache/paimon/fileindex/rangebitmap/RangeBitmapFileIndex.java
@@ -60,7 +60,7 @@ public class RangeBitmapFileIndex implements FileIndexer {
     }
 
     @Override
-    public FileIndexReader createReader(SeekableInputStream in, long start, int length) {
+    public FileIndexReader createReader(SeekableInputStream in, long start, long length) {
         return new Reader(dataType, in, start);
     }
